chore: sub header consistency
